Refactor upload component

+++ src/components/Upload/index.tsx
@@ -10,6 +10,7 @@ import React, {
 } from 'react';
 import FileUpload from './FileUpload';
 import './style.scss';
+import { AxiosProgressEvent } from 'axios';
 
 interface UploadProps {
   accept?: string;
@@ -17,7 +18,12 @@ interface UploadProps {
   files?: Partial<IFile>[];
   children?: ReactNode;
   round?: boolean;
+  showList?: boolean;
   onChange?: (files: IFile[]) => void;
+}
+
+export interface IProgress extends AxiosProgressEvent {
+  loading?: boolean;
 }
 
 const Upload: React.FC<UploadProps> = ({
@@ -25,6 +31,7 @@ const Upload: React.FC<UploadProps> = ({
   multiple,
   files = [],
   round,
+  showList = true,
   children,
   onChange,
 }) => {
@@ -32,11 +39,86 @@ const Upload: React.FC<UploadProps> = ({
   const ref = useRef<HTMLInputElement>(null);
 
   const [fileList, setFileList] = useState<IFile[]>(files as IFile[]);
+  const [fileInfo, setFileInfo] = useState<IFile[]>(files as IFile[]);
+  const [progress, setProgress] = useState<IProgress[]>([]);
+  const [errros, setErrors] = useState<string[]>([]);
   const [isDragOver, setIsDragOver] = useState(false);
+
+  const updateProgress = ({
+    progress,
+    index,
+    loading,
+  }: {
+    progress?: AxiosProgressEvent;
+    index: number;
+    loading: boolean;
+  }) => {
+    setProgress((p) => {
+      const newList = [...p];
+      if (progress) {
+        newList[index] = { ...progress, loading };
+      } else {
+        newList[index] = { ...p[index], loading };
+      }
+
+      return newList;
+    });
+  };
+
+  const handleUpload = (files: File[]) => {
+    const startIndex = multiple ? fileInfo.length : 0;
+
+    files.map((file, index) => {
+      uploadFile(file, startIndex + index);
+    });
+  };
+
+  const updateFile = (file: IFile, index: number) => {
+    const newFiles = [...fileInfo];
+    newFiles[index] = { ...file };
+
+    if (onChange) {
+      onChange(newFiles);
+    }
+
+    setFileInfo(newFiles);
+
+    if (onChange) {
+      onChange(newFiles);
+    }
+  };
+
+  const updateError = (error: string, index: number) => {
+    const errs = [...errros];
+    errs[index] = error;
+    setErrors(errs);
+  };
+
+  const uploadFile = (file: File, index: number) => {
+    updateError('', index);
+
+    return upload(file, (progress) =>
+      updateProgress({ progress, index, loading: true })
+    )
+      .then((uploaded) => {
+        updateFile(uploaded, index);
+      })
+      .catch((error) => {
+        updateError(error.message, index);
+      })
+      .finally(() => {
+        updateProgress({ index, loading: false });
+      });
+  };
 
   const updateFileList = (files: FileList) => {
     const newFiles = Array.from(files);
-    setFileList(fileList.concat(newFiles));
+    const newFilelist: IFile[] = multiple
+      ? [...fileList, ...newFiles]
+      : newFiles;
+
+    setFileList(newFilelist);
+    handleUpload(newFiles);
   };
 
   const clearValue = () => {
@@ -52,35 +134,23 @@ const Upload: React.FC<UploadProps> = ({
     }
   };
 
-  const handleFileChange = (file: IFile, index: number) => {
-    setFileList((list) => {
-      const newFiles = [...list];
-
-      newFiles[index] = file;
-
-      if (onChange) {
-        onChange(newFiles);
-      }
-
-      return newFiles;
-    });
-  };
-
   const removeFile = (index: number) => {
     const newFiles = [...fileList];
-    newFiles.splice(index, 1);
+    const newFileInfo = [...fileInfo];
 
     newFiles.splice(index, 1);
+    newFileInfo.splice(index, 1);
 
     setFileList(newFiles);
+    setFileInfo(newFileInfo);
 
     if (onChange) {
-      onChange(newFiles);
+      onChange(newFileInfo);
     }
   };
 
   const openFilePicker = () => {
-    if (ref.current && (multiple || !fileList.length)) {
+    if (ref.current) {
       clearValue();
 
       ref.current.click();
@@ -115,21 +185,22 @@ const Upload: React.FC<UploadProps> = ({
   };
 
   const renderFileList = () => {
-    if (fileList?.length) {
+    if (fileList?.length && showList) {
       return fileList.map((file, index) => (
         <FileUpload
           file={file}
           key={index}
-          upload={upload}
+          upload={() => uploadFile(file, index)}
           round={round}
+          progress={progress[index]}
           onRemove={() => removeFile(index)}
-          onChange={(fileInfo) => handleFileChange(fileInfo, index)}
+          error={errros[index]}
         />
       ));
     }
   };
 
-  const showPlaceholder = multiple || fileList.length < 1;
+  const showPlaceholder = multiple || fileList.length < 1 || !showList;
 
   return (
     <div className='lanting-upload'>
@@ -141,27 +212,29 @@ const Upload: React.FC<UploadProps> = ({
         ref={ref}
         multiple={multiple}
       />
-      <div className='lanting-upload-preview-list'>{renderFileList()}</div>
-      {showPlaceholder && (
-        <div
-          onClick={openFilePicker}
-          onDrop={handleDrop}
-          onDragOver={handleDragOver}
-          onDragLeave={handleDragLeave}
-          className='lanting-upload-trigger'
-        >
-          {children || (
-            <div
-              className={cls('lanting-upload-placeholder', {
-                'lanting-upload-placeholder--over': isDragOver,
-                'lanting-upload-placeholder--round': round,
-              })}
-            >
-              <Icon name='file_upload' />
-            </div>
-          )}
-        </div>
-      )}
+      <div className='lanting-upload-preview-list'>
+        {renderFileList()}
+        {showPlaceholder && (
+          <div
+            onClick={openFilePicker}
+            onDrop={handleDrop}
+            onDragOver={handleDragOver}
+            onDragLeave={handleDragLeave}
+            className='lanting-upload-trigger'
+          >
+            {children || (
+              <div
+                className={cls('lanting-upload-placeholder', {
+                  'lanting-upload-placeholder--over': isDragOver,
+                  'lanting-upload-placeholder--round': round,
+                })}
+              >
+                <Icon name='file_upload' />
+              </div>
+            )}
+          </div>
+        )}
+      </div>
     </div>
   );
 };
